Add check to make sure all tiles were classified.
I noticed that sometimes Feature Analyst was inexplicably failing to classify all tiles in the inputs folder. I added a check which verifies that all tif files in the Inputs folder are present in the Outputs folder. If not, an I/O error is raised and returns the missing file names.