Add object to check if gaps are present in mosaic
Couple of issues faced. 1. ArcPy will not load in a full region mosaic as an array. Due to this the method was only tested on a 2 tile mosaic. 2. Due to the iterative nature and large array size, it is slow. I made it so it is only iterating over only nodata cells as opposed to *every* cell, but with the equal area projection there is still a lot of nodata cells along the edges making it slow still.